updating-route-service

+++ ProductService.Infrastructure/Services/ImageService.cs
@@ -19,25 +19,42 @@ namespace ProductService.Infrastructure.Services
             if (!IsValidImage(fileName))
                 throw new ArgumentException("Invalid image format");
 
-            var uniqueFileName = $"{inventoryCode}-{DateTime.UtcNow.Ticks}{Path.GetExtension(fileName)}";
-            var filePath = Path.Combine(_imagePath, uniqueFileName);
+            var inventoryFolder=Path.Combine(_imagePath,inventoryCode.ToString());
+            Directory.CreateDirectory(inventoryFolder);
+
+            var uniqueFileName = $"{DateTime.UtcNow.Ticks}{Path.GetExtension(fileName)}";
+            var filePath = Path.Combine(inventoryFolder, uniqueFileName);
 
             using var fileStream = new FileStream(filePath, FileMode.Create);
             await imageStream.CopyToAsync(fileStream);
 
-            return $"/images/products/{uniqueFileName}";
+            return $"/images/products/{inventoryCode}/{uniqueFileName}";
         }
 
         public Task DeleteImageAsync(string imageUrl)
         {
             if (string.IsNullOrEmpty(imageUrl))
                 return Task.CompletedTask;
+            
+            var segments= imageUrl.Split('/');
+            if (segments.Length>=2)
+            {
+                var inventoryCode = segments[^2];
+                var fileName = segments[^1];
+                var filePath = Path.Combine(_imagePath, inventoryCode);
 
-            var fileName = Path.GetFileName(imageUrl);
-            var filePath = Path.Combine(_imagePath, fileName);
+                if(File.Exists(filePath))
+                    File.Delete(filePath);
+            }
+            return Task.CompletedTask;
+        }
 
-            if (File.Exists(filePath))
-                File.Delete(filePath);
+        public Task DeleteInventoryFolderAsync(int inventoryCode)
+        {
+            var folderPath = Path.Combine(_imagePath, inventoryCode.ToString());
+
+            if (Directory.Exists(folderPath))
+                Directory.Delete(folderPath, true); // true = recursive delete
 
             return Task.CompletedTask;
         }
